Featura/93653 - Ajuste nomenclatura de propriedades e componente genérico do histórico
