+++ src/@legacy/componentes-sgp/historico/historico.js
@@ -1,30 +1,20 @@
-import { Base, CardCollapse } from '@/@legacy/componentes';
-import { SGP_COLLAPSE_HISTORICO_ALTERACOES_NAAPA } from '@/@legacy/constantes/ids/collapse';
 import { HistoricoItem } from './historico-item';
 import { Paginacao } from '@/@legacy/componentes-sgp';
 import { useEffect, useState } from 'react';
-import ServicoNAAPA from '@/@legacy/servicos/Paginas/Gestao/NAAPA/ServicoNAAPA';
-import { useParams } from 'react-router-dom';
-import { erros } from '@/@legacy/servicos';
+import { api, erros } from '@/@legacy/servicos';
 
-export const HistoricoNAAPA = () => {
-  const { id: encaminhamentoNAAPAId } = useParams();
-
+export const HistoricoPaginado = ({ url }) => {
+  const [paginaAtual, setPaginaAtual] = useState(1);
   const [dadosHistorico, setDadosHistorico] = useState();
   const [numeroRegistros, setNumeroRegistros] = useState(1);
   const [numeroRegistrosPagina, setNumeroRegistrosPagina] = useState(10);
 
-  const configCabecalho = {
-    altura: '44px',
-    corBorda: Base.AzulBordaCollapse,
-  };
-
-  const obterHistorico = async (paginaAtual, numeroPag) => {
-    const resposta = await ServicoNAAPA.obterHistoricoPaginado(
-      encaminhamentoNAAPAId,
-      paginaAtual,
-      numeroPag
-    ).catch(e => erros(e));
+  const obterHistorico = async (numeroPagina, numeroRegistros) => {
+    const resposta = await api
+      .get(
+        `${url}?numeroPagina=${numeroPagina}&numeroRegistros=${numeroRegistros}`
+      )
+      .catch(e => erros(e));
 
     if (resposta?.data?.items) {
       setDadosHistorico(resposta.data.items);
@@ -35,7 +25,8 @@ export const HistoricoNAAPA = () => {
   };
 
   const onChangePaginacao = async pagina => {
-    obterHistorico(pagina, numeroRegistrosPagina, numeroRegistrosPagina);
+    setPaginaAtual(pagina);
+    obterHistorico(pagina, numeroRegistrosPagina);
   };
 
   const onChangeNumeroLinhas = async (paginaAtual, numeroLinhas) => {
@@ -45,21 +36,11 @@ export const HistoricoNAAPA = () => {
   };
 
   useEffect(() => {
-    if (encaminhamentoNAAPAId) {
-      obterHistorico(numeroRegistros, numeroRegistrosPagina);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [encaminhamentoNAAPAId]);
+    obterHistorico(paginaAtual, numeroRegistrosPagina);
+  }, []);
 
   return (
-    <CardCollapse
-      alt={`history-alt`}
-      key={`history-collapse-key`}
-      titulo="Histórico de alterações"
-      configCabecalho={configCabecalho}
-      indice={`history-collapse-indice`}
-      id={SGP_COLLAPSE_HISTORICO_ALTERACOES_NAAPA}
-    >
+    <>
       {dadosHistorico?.map((dados, index) => {
         return <HistoricoItem key={index} historico={dados} />;
       })}
@@ -73,6 +54,6 @@ export const HistoricoNAAPA = () => {
         onChangeNumeroLinhas={onChangeNumeroLinhas}
         pageSizeOptions={['10', '20', '50', '100']}
       />
-    </CardCollapse>
+    </>
   );
 };
