+++ src/@legacy/constantes/ids/collapse/index.js
@@ -25,6 +25,6 @@ export const SGP_COLLAPSE_LOCALIZAR_ESTUDANTE_CRIANCA =
 export const SGP_COLLAPSE_OBSERVACOES_PLANO_AEE =
   'SGP_COLLAPSE_OBSERVACOES_PLANO_AEE';
 export const SGP_COLLAPSE_PARECER_PLANO_AEE = 'SGP_COLLAPSE_PARECER_PLANO_AEE';
-export const SGP_COLLAPSE_HISTORICO_ALTERACOES_NAAPA =
-  'SGP_COLLAPSE_HISTORICO_ALTERACOES_NAAPA';
-export const SGP_COLLAPSE_OBSERVACOES_NAAPA = 'SGP_COLLAPSE_OBSERVACOES_NAAPA';
+export const SGP_COLLAPSE_HISTORICO_ALTERACOES =
+  'SGP_COLLAPSE_HISTORICO_ALTERACOES';
+export const SGP_COLLAPSE_OBSERVACOES = 'SGP_COLLAPSE_OBSERVACOES';
